Add gradle; add genetic algorithm training with jenetics

+++ src/main/java/IGameStateUtilityFunction.java
@@ -1,5 +1,5 @@
 import java.util.*;
 
 public interface IGameStateUtilityFunction {
-	public float get(GameState gameState);
+	public double get(GameState gameState);
 }
